backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0
pydantic==2.7.1
pydantic-settings==2.2.1
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.1
websockets==12.0
openai==1.30.1
google-generativeai==0.5.2
elevenlabs==0.2.27
pillow==10.3.0
numpy==1.26.4
